Added a slightly more informative progress bar
Fixed 2 warnings that could of caused compiling to fail with stricter compilers.

+++ lib/include/prism/core/result_types.h
@@ -8,15 +8,15 @@ namespace prism {
 namespace core {
 
 struct ArchiveCreationResult {
-    int files_added;
+    long files_added;
     uint64_t total_uncompressed_size;
     uint64_t total_compressed_size;
     std::vector<long long> thread_durations_ms;
 };
 
 struct ArchiveExtractionResult {
-    int files_extracted;
-    int files_skipped;
+    long files_extracted;
+    long files_skipped;
     uint64_t bytes_extracted;
     int hashes_checked;
     int hash_mismatches;
